+++ apps/www/pages/contributors/index.tsx
@@ -4,6 +4,7 @@ import type { Compressed } from "compress-json"
 import type { NextPage } from "next"
 import { NextSeo } from "next-seo"
 import Link from "next/link"
+import customEvents from "~/analytics/customEvents"
 import PageLayout, { Props as PageLayoutProps } from "~/pages/PageLayout"
 import getContributorHRef from "~/routes/getContributorHRef"
 import createListStaticPropsGetter from "~/ssg/createListStaticPropsGetter"
@@ -22,7 +23,10 @@ const PageComponent: NextPage<Props> = ({ fallback, ...props }) => (
                 description="A list of everyone who has contributed free silhouette images to PhyloPic."
                 title="Contributors to PhyloPic"
             />
-            <PaginationContainer endpoint={process.env.NEXT_PUBLIC_API_URL + "/contributors"}>
+            <PaginationContainer
+                endpoint={process.env.NEXT_PUBLIC_API_URL + "/contributors"}
+                onPage={index => customEvents.loadContributorListPage("contributors", index)}
+            >
                 {(contributors: readonly Contributor[], totalContributors: number) => (
                     <>
                         <header>
@@ -36,12 +40,16 @@ const PageComponent: NextPage<Props> = ({ fallback, ...props }) => (
                             </p>
                         </header>
                         <Board
-                            items={(contributors as readonly Contributor[]).map(({ _links, count, name, uuid }) => [
-                                uuid,
-                                <Link key={`link:${uuid}`} href={getContributorHRef(_links.self)}>
-                                    {name || "Anonymous"}
+                            items={(contributors as readonly Contributor[]).map(contributor => [
+                                contributor.uuid,
+                                <Link
+                                    key={`link:${contributor.uuid}`}
+                                    href={getContributorHRef(contributor._links.self)}
+                                    onClick={() => customEvents.clickContributorLink("contributor_board", contributor)}
+                                >
+                                    {contributor.name || "Anonymous"}
                                 </Link>,
-                                <NumberView key={`count:${uuid}`} value={count} />,
+                                <NumberView key={`count:${contributor.uuid}`} value={contributor.count} />,
                             ])}
                         />
                     </>