+++ apps/www/pages/donate/cancel/index.tsx
@@ -1,15 +1,18 @@
 import type { NextPage } from "next"
 import { NextSeo } from "next-seo"
-import Link from "next/link"
+import customEvents from "~/analytics/customEvents"
 import PageLayout from "~/pages/PageLayout"
 import Breadcrumbs from "~/ui/Breadcrumbs"
-import BulletList from "~/ui/BulletList"
+import EngineeringSection from "~/sections/EngineeringSection"
 import InlineSections from "~/ui/InlineSections"
+import PatreonSection from "~/sections/PatreonSection"
 import SiteTitle from "~/ui/SiteTitle"
+import SpreadWordSection from "~/sections/SpreadWordSection"
+import TechDonateList from "~/sections/TechDonateList"
 const PageComponent: NextPage = () => (
     <PageLayout>
         <NextSeo
-            canonical={`${process.env.NEXT_PUBLIC_WWW_URL}/contribute/cancel`}
+            canonical={`${process.env.NEXT_PUBLIC_WWW_URL}/donate/cancel`}
             description="Ways to contribute to PhyloPic, an open database of freely reusable silhouette images of organisms."
             noindex
             title="Other Ways to Contribute to PhyloPic"
@@ -18,7 +21,7 @@ const PageComponent: NextPage = () => (
             <Breadcrumbs
                 items={[
                     { children: "Home", href: "/" },
-                    { children: "Donate", href: "https://www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW" },
+                    { children: "Donate", href: "/donate" },
                     { children: <strong>Other Ways to Contribute</strong> },
                 ]}
             />
@@ -40,74 +43,37 @@ const PageComponent: NextPage = () => (
                 </p>
                 <p></p>
             </section>
-            <section>
-                <h2>Become a Patron</h2>
-                <p>
-                    For as little as $1 a month, you can see previews of new <SiteTitle /> functionality, as well as
-                    updates on other projects by{" "}
-                    <Link href={`/contributors/${process.env.NEXT_PUBLIC_CONTACT_CONTRIBUTOR_UUID}`}>Mike Keesey</Link>,
-                    like the comic book series{" "}
-                    <a href="//www.keesey-comics.com/paleocene" rel="external">
-                        <cite>Paleocene</cite>
-                    </a>
-                    .{" "}
-                    <a href="//www.patreon.com/tmkeesey?fan_landing=true" rel="author">
-                        Become a patron!
-                    </a>
-                </p>
-            </section>
-            <section>
-                <h2>Spread the Word!</h2>
-                <p>
-                    Tell people about{" "}
-                    <a href={`${process.env.NEXT_PUBLIC_WWW_URL}`}>
-                        <SiteTitle />
-                    </a>
-                    !
-                </p>
-            </section>
+            <PatreonSection />
+            <SpreadWordSection />
             <section>
                 <h2>Support Free Technologies</h2>
                 <p>
                     <SiteTitle /> relies on a number of free, open-source technologies. Support them instead! Here are a
                     few:
                 </p>
-                <BulletList>
-                    <li>
-                        <a href="//inkscape.org/support-us/" rel="external">
-                            Inkscape
-                        </a>
-                    </li>
-                    <li>
-                        <a href="//imagemagick.org/script/support.php#support" rel="external">
-                            ImageMagick
-                        </a>
-                    </li>
-                    <li>
-                        <a href="//opencollective.com/mochajs#support" rel="external">
-                            Mocha
-                        </a>
-                    </li>
-                </BulletList>
+                <TechDonateList />
             </section>
-            <section>
-                <h2>Software Engineering</h2>
-                <p>
-                    If you are technically inclined, check out the{" "}
-                    <a href="//github.com/keesey/phylopic">code repository</a> and/or the{" "}
-                    <a href="http://api-docs.phylopic.org/v2" rel="help">
-                        API Documentation
-                    </a>
-                    . Think about contributing to <SiteTitle /> or building a tool that uses it.
-                </p>
-            </section>
+            <EngineeringSection />
             <section>
                 <h2>Change Your Mind?</h2>
                 <p>
-                    Come on, you want to donate <em>something</em> don’t you? Right?
+                    Come on, you want to donate <em>something</em> don&rsquo;t you? Right?
                 </p>
                 <p>
-                    <Link href="/donate">Go back!</Link>
+                    <a
+                        href="//www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW"
+                        onClick={() =>
+                            customEvents.clickLink(
+                                "donate_cancel_go_back",
+                                "//www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW",
+                                "Go Back!",
+                                "link",
+                            )
+                        }
+                        rel="noreferrer"
+                    >
+                        Go back!
+                    </a>
                 </p>
             </section>
         </InlineSections>