+++ apps/www/src/views/ContributionCTAView/index.tsx
@@ -1,5 +1,6 @@
 import Link from "next/link"
 import { FC } from "react"
+import customEvents from "~/analytics/customEvents"
 import SiteTitle from "~/ui/SiteTitle"
 import styles from "./index.module.scss"
 const ContributionCTAView: FC = () => {
@@ -14,7 +15,19 @@ const ContributionCTAView: FC = () => {
                     </p>
                     <ul className={styles.ctaList}>
                         <li>
-                            <a href={process.env.NEXT_PUBLIC_CONTRIBUTE_URL + "/"}>Image Uploader</a>
+                            <a
+                                href={process.env.NEXT_PUBLIC_CONTRIBUTE_URL + "/"}
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "contribution_cta_link",
+                                        process.env.NEXT_PUBLIC_CONTRIBUTE_URL + "/",
+                                        "Image Uplaoder",
+                                        "link",
+                                    )
+                                }
+                            >
+                                Image Uploader
+                            </a>
                         </li>
                     </ul>
                 </section>
@@ -27,15 +40,50 @@ const ContributionCTAView: FC = () => {
                     </p>
                     <ul className={styles.ctaList}>
                         <li>
-                            <Link href="/articles/api-recipes">API Recipes</Link>
+                            <Link
+                                href="/articles/api-recipes"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "contribution_cta_link",
+                                        "/articles/api-recipes",
+                                        "API Recipes",
+                                        "link",
+                                    )
+                                }
+                            >
+                                API Recipes
+                            </Link>
                         </li>
                         <li>
-                            <a href="http://api-docs.phylopic.org/v2" rel="help">
+                            <a
+                                href="http://api-docs.phylopic.org/v2"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "contribution_cta_link",
+                                        "http://api-docs.phylopic.org/v2",
+                                        "API Documentation",
+                                        "link",
+                                    )
+                                }
+                                rel="help"
+                            >
                                 API Documentation
                             </a>
                         </li>
                         <li>
-                            <a href="//github.com/keesey/phylopic">Codebase</a>
+                            <a
+                                href="//github.com/keesey/phylopic"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "//github.com/keesey/phylopic",
+                                        "http://api-docs.phylopic.org/v2",
+                                        "Codebase",
+                                        "link",
+                                    )
+                                }
+                            >
+                                Codebase
+                            </a>
                         </li>
                     </ul>
                 </section>
@@ -47,11 +95,49 @@ const ContributionCTAView: FC = () => {
                     </p>
                     <ul className={styles.ctaList}>
                         <li>
-                            <a href="//www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW">Make a donation</a>
+                            <a
+                                href="//www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "//github.com/keesey/phylopic",
+                                        "//www.paypal.com/donate/?hosted_button_id=9GL697FDK7ZWW",
+                                        "Make a donation",
+                                        "link",
+                                    )
+                                }
+                            >
+                                Make a donation
+                            </a>
                         </li>
                         <li>
-                            <a href="//www.patreon.com/tmkeesey?fan_landing=true" rel="author">
+                            <a
+                                href="//www.patreon.com/tmkeesey?fan_landing=true"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "//github.com/keesey/phylopic",
+                                        "//www.patreon.com/tmkeesey?fan_landing=true",
+                                        "Support the creator on Patreon",
+                                        "link",
+                                    )
+                                }
+                                rel="author"
+                            >
                                 Support the creator on Patreon
+                            </a>
+                        </li>
+                        <li>
+                            <a
+                                href="//www.buymeacoffee.com/phylopic"
+                                onClick={() =>
+                                    customEvents.clickLink(
+                                        "menu_link",
+                                        "//www.buymeacoffee.com/phylopic",
+                                        "Buy me a coffee",
+                                        "link",
+                                    )
+                                }
+                            >
+                                Buy me a coffee
                             </a>
                         </li>
                     </ul>