+++ apps/www/src/views/LinkedImageThumbnailView/index.tsx
@@ -4,6 +4,7 @@ import { extractPath, stringifyNormalized } from "@phylopic/utils"
 import clsx from "clsx"
 import Link from "next/link"
 import { DragEvent, FC } from "react"
+import customEvents from "~/analytics/customEvents"
 import styles from "./index.module.scss"
 export interface Props {
     inverted?: boolean
@@ -11,6 +12,7 @@ export interface Props {
 }
 const LinkedImageThumbnailView: FC<Props> = ({ inverted, value }) => {
     const handleDragStart = (event: DragEvent) => {
+        customEvents.dragImage("thumbnail", value)
         event.dataTransfer.setData(DATA_MEDIA_TYPE, stringifyNormalized(value))
     }
     return (
@@ -18,6 +20,7 @@ const LinkedImageThumbnailView: FC<Props> = ({ inverted, value }) => {
             className={clsx(styles.main, !inverted && styles.standard, inverted && styles.inverted)}
             draggable
             href={extractPath(value._links.self.href)}
+            onClick={() => customEvents.clickImageLink("thumbnail", value, undefined, "button")}
             onDragStart={handleDragStart}
             title={value._links.self.title}
         >