+++ packages/ui/CHANGELOG.md
@@ -9,11 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Analytics code (GA4).
+-   `onPage()` property for `PaginationContainer`.
+
 ### Changed
 
 ### Deprecated
 
 ### Fixed
+
+-   Missing dependency in `useImageLoader()`.
+-   Alignment of `InfiniteScroll` loader.
 
 ### Removed